requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.5
Flask-Bootstrap==3.3.7.1
WTForms==2.3.1
Flask-WTF==0.14.3
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Mail==0.9.1
email-validator==1.0.5
